modbus-meter-sim/requirements.txt
pymodbus==3.*
pyserial>=3.5,<4
